[Presentation] Removed the tabs in the banner history page
