+++ lib/application/banner_history/banner_history_event.dart
@@ -1,0 +1,8 @@
+part of 'banner_history_bloc.dart';
+
+@freezed
+class BannerHistoryEvent with _$BannerHistoryEvent {
+  const factory BannerHistoryEvent.init({
+    @Default(BannerHistoryItemType.character) BannerHistoryItemType type,
+  }) = _Init;
+}
