+++ lib/application/banner_history/banner_history_state.dart
@@ -1,0 +1,9 @@
+part of 'banner_history_bloc.dart';
+
+@freezed
+class BannerHistoryState with _$BannerHistoryState {
+  const factory BannerHistoryState.initial({
+    required List<BannerHistoryItemModel> banners,
+    required List<double> versions,
+  }) = _InitialState;
+}
