--- lib/application/banner_history/charts/banner_history_charts_event.dart
@@ -1,6 +1,0 @@
-part of 'banner_history_charts_bloc.dart';
-
-@freezed
-class BannerHistoryChartsEvent with _$BannerHistoryChartsEvent {
-  const factory BannerHistoryChartsEvent.init() = _Init;
-}
