--- lib/application/banner_history/charts/banner_history_charts_state.dart
@@ -1,6 +1,0 @@
-part of 'banner_history_charts_bloc.dart';
-
-@freezed
-class BannerHistoryChartsState with _$BannerHistoryChartsState {
-  const factory BannerHistoryChartsState.initial() = _InitialState;
-}
