--- lib/application/banner_history/versions/banner_history_versions_event.dart
@@ -1,8 +1,0 @@
-part of 'banner_history_versions_bloc.dart';
-
-@freezed
-class BannerHistoryVersionsEvent with _$BannerHistoryVersionsEvent {
-  const factory BannerHistoryVersionsEvent.init({
-    @Default(BannerHistoryItemType.character) BannerHistoryItemType type,
-  }) = _Init;
-}
